Add tfcli.Client.DiscardOperation to discard results
