+++ go.mod
@@ -4,7 +4,6 @@ go 1.16
 
 require (
 	github.com/crossplane/crossplane-runtime v0.14.0
-	github.com/google/uuid v1.1.2
 	github.com/hashicorp/terraform-plugin-sdk/v2 v2.7.0
 	github.com/iancoleman/strcase v0.2.0
 	github.com/json-iterator/go v1.1.10
